Added Sotheryn's fuel price approximation logic to the bot

+++ FuelInfo.cs
@@ -4,9 +4,9 @@ namespace TauManagerBot
 {
         public class FuelInfo
         {
-            public double Last_Price { get; set; }
-            public double Max_Price { get; set; }
-            public double Min_Price { get; set; }
+            public decimal Last_Price { get; set; }
+            public decimal Max_Price { get; set; }
+            public decimal Min_Price { get; set; }
             public DateTime Last_Reading { get; set; }
             public string Station_Name { get; set; }
             public string Station_Short_Name { get; set; }
